requirements.txt
asgiref==3.7.2
Django==4.2.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Pillow==10.0.0
PyJWT==2.9.0
sqlparse==0.4.4
typing_extensions==4.7.1
